+++ lib/src/extensions/duration.dart
@@ -12,8 +12,8 @@ extension DurationExt on Duration {
   ///   }
   ///
   ///```
-  Future<T> delay<T>([FutureOr<T> Function()? callback]) async =>
-      Future<T>.delayed(this, callback);
+  Future<T> delayed<T>([FutureOr<T> Function()? computation]) async =>
+      Future<T>.delayed(this, computation);
 
   /// Adds the Duration to the current DateTime and returns a DateTime in the future
   DateTime get fromNow => DateTime.now() + this;